Avanzando en la implementación de múltiples entornos
git-svn-id: https://forja.rediris.es/svn/cusl4-catastrof@136 8c4acde5-3c00-4ed3-8754-bc08a9928e99

+++ googleEarth/GoogleEarthUtils.java
@@ -39,9 +39,9 @@ public class GoogleEarthUtils {
 	/**
 	 * Size in meters of the circunflex circle of the hexagon
 	 */
-	protected static short tileSize;
+	protected static int tileSize;
 
-	public static void setTileSize(short tileSize) {
+	public static void setTileSize(int tileSize) {
 		GoogleEarthUtils.tileSize = tileSize;
 	}
